feat(dapp): Move policies from dialogs to pages
