+++ dapp/src/app/layout.tsx
@@ -8,7 +8,6 @@ import { Suspense } from 'react';
 
 import { ConnectionGuard } from '@/components';
 import { AvailabilityCheckDialog } from '@/components/availability-check/AvailabilityCheckDialog';
-import { TermsAndConditionsDialog } from '@/components/dialogs/TermsAndConditionsDialog';
 import { Footer, Navbar } from '@/components/layout';
 import { DEFAULT_METADATA } from '@/lib/constants/metadata.constants';
 import { APP_STATIC_THEME } from '@/lib/constants/theme.constants';
@@ -31,7 +30,6 @@ export default function RootLayout({
                             {children}
                             <AvailabilityCheckDialog />
                             <Footer />
-                            <TermsAndConditionsDialog />
                         </ConnectionGuard>
                     </Suspense>
                 </AppProviders>
