+++ dapp/src/legal/tos.mdx
@@ -87,7 +87,7 @@ IOTA NAMES DISCLAIMS ANY LIABILITY FOR DIRECT, INDIRECT, INCIDENTAL, SPECIAL, CO
 
 The Service operates on a decentralized network. All on-chain interactions and name registrations occur publicly and are permanently recorded on the IOTA ledger. While the Service itself does not actively collect or store personal data off-chain, certain information may, in specific contexts, qualify as pseudonymous personal data if linked to an identifiable individual through additional information.
 
-We may also process limited off-chain technical or analytics data as described in our <span className="font-medium underline" > [Privacy Policy ](?modal=privacy_policy) </span>. By using the Service, you acknowledge and agree to the processing of data as outlined therein.
+We may also process limited off-chain technical or analytics data as described in our <span className="font-medium underline" > [Privacy Policy ](/privacy) </span>. By using the Service, you acknowledge and agree to the processing of data as outlined therein.
 
 **11\. Changes to Terms**
