+++ dapp/src/lib/constants/routes.constants.tsx
@@ -46,10 +46,10 @@ export const FOOTER_SOCIAL_LINKS: Route[] = [
 export const FOOTER_LEGAL_LINKS: Route[] = [
     {
         title: 'Terms & Conditions',
-        path: '?modal=terms_conditions',
+        path: '/tos',
     },
     {
         title: 'Privacy Policy',
-        path: '?modal=privacy_policy',
+        path: '/privacy',
     },
 ];
