--- dapp/src/lib/utils/termsAndConditions.ts
@@ -1,22 +1,0 @@
-// Copyright (c) 2025 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-export class TermsAndConds {
-    static areAccepted(): boolean {
-        if (typeof window === 'undefined') {
-            return false;
-        }
-        try {
-            return JSON.parse(
-                window.localStorage.getItem('terms-and-conditions-accepted') || 'false',
-            );
-        } catch {
-            return false;
-        }
-    }
-
-    static accept() {
-        if (typeof window === 'undefined') return;
-        window.localStorage.setItem('terms-and-conditions-accepted', 'true');
-    }
-}
